feat: switch to runtime adapter and clean the code
