+++ packages/plugin-rag/src/index.ts
@@ -9,26 +9,6 @@ import { logger } from '@elizaos/core';
 import { validateModelConfig } from './config';
 import { RagService } from './service';
 
-// Export main functions and types
-export { generateText, generateTextEmbedding } from './llm';
-export { validateModelConfig, getProviderRateLimits } from './config';
-export type { ModelConfig, ProviderRateLimits, TextGenerationOptions } from './types';
-
-// Export RagService
-export { RagService } from './service';
-
-// Export contextual embedding utilities
-export {
-  DEFAULT_CHUNK_TOKEN_SIZE,
-  DEFAULT_CHUNK_OVERLAP_TOKENS,
-  CONTEXT_TARGETS,
-  getContextualizationPrompt,
-  getCachingContextualizationPrompt,
-  getPromptForMimeType,
-  getCachingPromptForMimeType,
-  getChunkWithContext,
-} from './ctx-embeddings';
-
 /**
  * RAG Plugin - Provides Retrieval Augmented Generation capabilities
  */
@@ -37,10 +17,13 @@ export const ragPlugin: Plugin = {
   description:
     'Plugin for Retrieval Augmented Generation, including knowledge management and embedding.',
   config: {
-    // Required env variables for the plugin
+    // Embedding configuration (can use plugin-openai settings automatically)
     EMBEDDING_PROVIDER: process.env.EMBEDDING_PROVIDER,
-    TEXT_PROVIDER: process.env.TEXT_PROVIDER,
     TEXT_EMBEDDING_MODEL: process.env.TEXT_EMBEDDING_MODEL,
+    EMBEDDING_DIMENSION: process.env.EMBEDDING_DIMENSION,
+
+    // Text generation (only required when CTX_RAG_ENABLED=true)
+    TEXT_PROVIDER: process.env.TEXT_PROVIDER,
     TEXT_MODEL: process.env.TEXT_MODEL,
 
     // API keys (will depend on selected providers)
@@ -76,9 +59,18 @@ export const ragPlugin: Plugin = {
           `Using ${validatedConfig.EMBEDDING_PROVIDER} for embeddings and ${validatedConfig.TEXT_PROVIDER} for text generation.`
         );
       } else {
-        logger.info(
-          'Running in Basic Embedding mode (CTX_RAG_ENABLED=false). TEXT_PROVIDER and TEXT_MODEL not required.'
-        );
+        const usingPluginOpenAI = !process.env.EMBEDDING_PROVIDER;
+
+        if (usingPluginOpenAI) {
+          logger.info(
+            'Running in Basic Embedding mode with auto-detected configuration from plugin-openai.'
+          );
+        } else {
+          logger.info(
+            'Running in Basic Embedding mode (CTX_RAG_ENABLED=false). TEXT_PROVIDER and TEXT_MODEL not required.'
+          );
+        }
+
         logger.info(
           `Using ${validatedConfig.EMBEDDING_PROVIDER} for embeddings with ${validatedConfig.TEXT_EMBEDDING_MODEL}.`
         );
@@ -87,9 +79,7 @@ export const ragPlugin: Plugin = {
       logger.info('Model configuration validated successfully.');
 
       if (runtime) {
-        logger.info(
-          `RAG Plugin global init for agent: ${runtime.agentId}. Per-agent RAG services will manage their own workers.`
-        );
+        logger.info(`RAG Plugin initialized for agent: ${runtime.agentId}`);
       }
       logger.info('RAG Plugin initialized.');
     } catch (error) {
